pyproject.toml
[project]
name = "metepdf"
version = "0.1.0"
description = "Convert TXT, images, DOCX, and PDF files to PDF from the command line."
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0.0",
    "reportlab>=4.0.0",
    "docx2pdf>=0.1.8",
    "PyMuPDF>=1.23.0",
    "customtkinter>=5.2.0",
    "tkinterdnd2>=0.3.0",
]

[project.scripts]
metepdf     = "main:main"
metepdf-gui = "gui:main"

[dependency-groups]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
